python-flask-server/swagger_server/controllers/default_controller.py
import connexion
import six

from swagger_server.models.api_anomalydetection_body import ApiAnomalydetectionBody  # noqa: E501
from swagger_server.models.api_dataenrichment_body import ApiDataenrichmentBody  # noqa: E501
from swagger_server.models.api_datatransformation_body import ApiDatatransformationBody  # noqa: E501
from swagger_server.models.api_habitats_body import ApiHabitatsBody  # noqa: E501
from swagger_server.models.api_temporalanalysis_body import ApiTemporalanalysisBody  # noqa: E501
from swagger_server.models.api_trainmodel_body import ApiTrainmodelBody  # noqa: E501
from swagger_server.models.augmentation_remotesensing_body import AugmentationRemotesensingBody  # noqa: E501
from swagger_server.models.caching_body import CachingBody  # noqa: E501
from swagger_server.models.dataquality_body import DataqualityBody  # noqa: E501
from swagger_server.models.error_response import ErrorResponse  # noqa: E501
from swagger_server.models.inline_response200 import InlineResponse200  # noqa: E501
from swagger_server.models.inline_response2001 import InlineResponse2001  # noqa: E501
from swagger_server.models.inline_response20010 import InlineResponse20010  # noqa: E501
from swagger_server.models.inline_response2002 import InlineResponse2002  # noqa: E501
from swagger_server.models.inline_response2003 import InlineResponse2003  # noqa: E501
from swagger_server.models.inline_response2004 import InlineResponse2004  # noqa: E501
from swagger_server.models.inline_response2005 import InlineResponse2005  # noqa: E501
from swagger_server.models.inline_response2006 import InlineResponse2006  # noqa: E501
from swagger_server.models.inline_response2007 import InlineResponse2007  # noqa: E501
from swagger_server.models.inline_response2008 import InlineResponse2008  # noqa: E501
from swagger_server.models.inline_response2009 import InlineResponse2009  # noqa: E501
from swagger_server.models.machine_learning_model import MachineLearningModel  # noqa: E501
from swagger_server.models.mapping_mapbox_body import MappingMapboxBody  # noqa: E501
from swagger_server.models.mlpreprocessing_body import MlpreprocessingBody  # noqa: E501
from swagger_server.models.rodent_habitat_detection import RodentHabitatDetection  # noqa: E501
from swagger_server import util


def adaptive_learning_post(body):  # noqa: E501
    """Enable or configure adaptive learning

    Configure adaptive learning parameters for continuous model improvements. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2005
    """
    if connexion.request.is_json:
        body = MachineLearningModel.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_anomaly_detection_post(body):  # noqa: E501
    """Detect unusual patterns in environmental data using anomaly detection models.

     # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2003
    """
    if connexion.request.is_json:
        body = ApiAnomalydetectionBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_data_enrichment_post(body):  # noqa: E501
    """Enrich datasets with external sources (geospatial or weather data).

     # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2003
    """
    if connexion.request.is_json:
        body = ApiDataenrichmentBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_data_transformation_post(body):  # noqa: E501
    """Transform datasets for machine learning purposes (pivoting, melting, aggregation).

     # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2003
    """
    if connexion.request.is_json:
        body = ApiDatatransformationBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_habitats_post(body):  # noqa: E501
    """Analyze satellite or environmental data to identify potential habitats for Mastomys Natalensis.

     # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse200
    """
    if connexion.request.is_json:
        body = ApiHabitatsBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_predict_movements_get(location, _date):  # noqa: E501
    """Forecast potential future movements of rodent populations.

     # noqa: E501

    :param location: Location for prediction
    :type location: dict | bytes
    :param _date: Date for prediction
    :type _date: dict | bytes

    :rtype: InlineResponse2002
    """
    if connexion.request.is_json:
        location = Object.from_dict(connexion.request.get_json())  # noqa: E501
    if connexion.request.is_json:
        _date = Object.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_temporal_analysis_post(body):  # noqa: E501
    """Analyze temporal data for population dynamics using models like LSTM or RNN.

     # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2003
    """
    if connexion.request.is_json:
        body = ApiTemporalanalysisBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def api_train_model_post(body):  # noqa: E501
    """Train machine learning models (CNN, SVM, Random Forest) using geospatial or satellite imagery.

     # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2001
    """
    if connexion.request.is_json:
        body = ApiTrainmodelBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def augmentation_remote_sensing_post(body):  # noqa: E501
    """Apply Augmentation to Remote Sensing Data

    Applies augmentation techniques to satellite imagery or LiDAR data to improve detection accuracy for Mastomys Natalensis habitat mapping. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2004
    """
    if connexion.request.is_json:
        body = AugmentationRemotesensingBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def caching_post(body):  # noqa: E501
    """Configure caching for performance optimization

    Set caching strategies to improve performance for repeated queries. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: None
    """
    if connexion.request.is_json:
        body = CachingBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def context_aware_post(body):  # noqa: E501
    """Enable or configure context-aware responses

    Set up context-aware response handling for dynamic and relevant interactions. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2006
    """
    if connexion.request.is_json:
        body = RodentHabitatDetection.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def data_quality_post(body):  # noqa: E501
    """Monitor and report data quality

    Checks data quality by identifying missing values, outliers, and other inconsistencies. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2007
    """
    if connexion.request.is_json:
        body = DataqualityBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def integrate_mapbox(body):  # noqa: E501
    """Integrate with Mapbox for interactive map visualization

    Provides integration with Mapbox to create and display interactive maps with geospatial data, including rodent habitat locations and environmental overlays. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2009
    """
    if connexion.request.is_json:
        body = MappingMapboxBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def integrate_openweather(latitude, longitude):  # noqa: E501
    """Retrieve real-time weather data from OpenWeather

    Fetches real-time weather data for specified location coordinates using the OpenWeather API. # noqa: E501

    :param latitude: 
    :type latitude: dict | bytes
    :param longitude: 
    :type longitude: dict | bytes

    :rtype: InlineResponse20010
    """
    if connexion.request.is_json:
        latitude = Object.from_dict(connexion.request.get_json())  # noqa: E501
    if connexion.request.is_json:
        longitude = Object.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def ml_preprocessing_post(body):  # noqa: E501
    """Preprocess data for ML

    Prepares data for ML, including encoding features and train/test splitting. # noqa: E501

    :param body: 
    :type body: dict | bytes

    :rtype: InlineResponse2008
    """
    if connexion.request.is_json:
        body = MlpreprocessingBody.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def rate_limiting_post():  # noqa: E501
    """Configure API rate limiting

    Set rate limits to avoid overloading the API with too many requests. # noqa: E501


    :rtype: None
    """
    return 'do some magic!'
