python-flask-server/swagger_server/test/test_default_controller.py
# coding: utf-8

from __future__ import absolute_import

from flask import json
from six import BytesIO

from swagger_server.models.api_anomalydetection_body import ApiAnomalydetectionBody  # noqa: E501
from swagger_server.models.api_dataenrichment_body import ApiDataenrichmentBody  # noqa: E501
from swagger_server.models.api_datatransformation_body import ApiDatatransformationBody  # noqa: E501
from swagger_server.models.api_habitats_body import ApiHabitatsBody  # noqa: E501
from swagger_server.models.api_temporalanalysis_body import ApiTemporalanalysisBody  # noqa: E501
from swagger_server.models.api_trainmodel_body import ApiTrainmodelBody  # noqa: E501
from swagger_server.models.augmentation_remotesensing_body import AugmentationRemotesensingBody  # noqa: E501
from swagger_server.models.caching_body import CachingBody  # noqa: E501
from swagger_server.models.dataquality_body import DataqualityBody  # noqa: E501
from swagger_server.models.error_response import ErrorResponse  # noqa: E501
from swagger_server.models.inline_response200 import InlineResponse200  # noqa: E501
from swagger_server.models.inline_response2001 import InlineResponse2001  # noqa: E501
from swagger_server.models.inline_response20010 import InlineResponse20010  # noqa: E501
from swagger_server.models.inline_response2002 import InlineResponse2002  # noqa: E501
from swagger_server.models.inline_response2003 import InlineResponse2003  # noqa: E501
from swagger_server.models.inline_response2004 import InlineResponse2004  # noqa: E501
from swagger_server.models.inline_response2005 import InlineResponse2005  # noqa: E501
from swagger_server.models.inline_response2006 import InlineResponse2006  # noqa: E501
from swagger_server.models.inline_response2007 import InlineResponse2007  # noqa: E501
from swagger_server.models.inline_response2008 import InlineResponse2008  # noqa: E501
from swagger_server.models.inline_response2009 import InlineResponse2009  # noqa: E501
from swagger_server.models.machine_learning_model import MachineLearningModel  # noqa: E501
from swagger_server.models.mapping_mapbox_body import MappingMapboxBody  # noqa: E501
from swagger_server.models.mlpreprocessing_body import MlpreprocessingBody  # noqa: E501
from swagger_server.models.rodent_habitat_detection import RodentHabitatDetection  # noqa: E501
from swagger_server.test import BaseTestCase


class TestDefaultController(BaseTestCase):
    """DefaultController integration test stubs"""

    def test_adaptive_learning_post(self):
        """Test case for adaptive_learning_post

        Enable or configure adaptive learning
        """
        body = MachineLearningModel()
        response = self.client.open(
            '/adaptive-learning',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_anomaly_detection_post(self):
        """Test case for api_anomaly_detection_post

        Detect unusual patterns in environmental data using anomaly detection models.
        """
        body = ApiAnomalydetectionBody()
        response = self.client.open(
            '/api/anomaly-detection',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_data_enrichment_post(self):
        """Test case for api_data_enrichment_post

        Enrich datasets with external sources (geospatial or weather data).
        """
        body = ApiDataenrichmentBody()
        response = self.client.open(
            '/api/data-enrichment',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_data_transformation_post(self):
        """Test case for api_data_transformation_post

        Transform datasets for machine learning purposes (pivoting, melting, aggregation).
        """
        body = ApiDatatransformationBody()
        response = self.client.open(
            '/api/data-transformation',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_habitats_post(self):
        """Test case for api_habitats_post

        Analyze satellite or environmental data to identify potential habitats for Mastomys Natalensis.
        """
        body = ApiHabitatsBody()
        response = self.client.open(
            '/api/habitats',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_predict_movements_get(self):
        """Test case for api_predict_movements_get

        Forecast potential future movements of rodent populations.
        """
        query_string = [('location', Object()),
                        ('_date', Object())]
        response = self.client.open(
            '/api/predict-movements',
            method='GET',
            query_string=query_string)
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_temporal_analysis_post(self):
        """Test case for api_temporal_analysis_post

        Analyze temporal data for population dynamics using models like LSTM or RNN.
        """
        body = ApiTemporalanalysisBody()
        response = self.client.open(
            '/api/temporal-analysis',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_api_train_model_post(self):
        """Test case for api_train_model_post

        Train machine learning models (CNN, SVM, Random Forest) using geospatial or satellite imagery.
        """
        body = ApiTrainmodelBody()
        response = self.client.open(
            '/api/train-model',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_augmentation_remote_sensing_post(self):
        """Test case for augmentation_remote_sensing_post

        Apply Augmentation to Remote Sensing Data
        """
        body = AugmentationRemotesensingBody()
        response = self.client.open(
            '/augmentation/remote-sensing',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_caching_post(self):
        """Test case for caching_post

        Configure caching for performance optimization
        """
        body = CachingBody()
        response = self.client.open(
            '/caching',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_context_aware_post(self):
        """Test case for context_aware_post

        Enable or configure context-aware responses
        """
        body = RodentHabitatDetection()
        response = self.client.open(
            '/context-aware',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_data_quality_post(self):
        """Test case for data_quality_post

        Monitor and report data quality
        """
        body = DataqualityBody()
        response = self.client.open(
            '/data-quality',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_integrate_mapbox(self):
        """Test case for integrate_mapbox

        Integrate with Mapbox for interactive map visualization
        """
        body = MappingMapboxBody()
        response = self.client.open(
            '/integration/mapping/mapbox',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_integrate_openweather(self):
        """Test case for integrate_openweather

        Retrieve real-time weather data from OpenWeather
        """
        query_string = [('latitude', Object()),
                        ('longitude', Object())]
        response = self.client.open(
            '/integration/weather/openweather',
            method='GET',
            query_string=query_string)
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_ml_preprocessing_post(self):
        """Test case for ml_preprocessing_post

        Preprocess data for ML
        """
        body = MlpreprocessingBody()
        response = self.client.open(
            '/ml-preprocessing',
            method='POST',
            data=json.dumps(body),
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))

    def test_rate_limiting_post(self):
        """Test case for rate_limiting_post

        Configure API rate limiting
        """
        response = self.client.open(
            '/rate-limiting',
            method='POST',
            content_type='application/json')
        self.assert200(response,
                       'Response body is : ' + response.data.decode('utf-8'))


if __name__ == '__main__':
    import unittest
    unittest.main()
